refactor(postgres): split restore into restore/ module (run, prepare, command, toc)
