+++ src/domain/postgres/restore/run.rs
@@ -1,122 +1,18 @@
 use anyhow::Result;
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::Command;
 use std::sync::Arc;
-use std::time::Instant;
 
-use super::clean_mode::RestoreCleanMode;
-use super::connection::{
+use super::{prepare_archive, run_pg_restore, toc_creates_public_schema};
+use crate::domain::postgres::clean_mode::RestoreCleanMode;
+use crate::domain::postgres::connection::{
     can_drop_database, drop_all_schemas, drop_and_recreate_database, pg_restore_binary_name,
     recreate_public_schema, select_pg_path, server_version, terminate_connections,
 };
-use super::format::PostgresDumpFormat;
+use crate::domain::postgres::format::PostgresDumpFormat;
 use crate::services::backup::logger::JobLogger;
 use crate::services::config::DatabaseConfig;
-
-pub(crate) fn toc_creates_public_schema(toc: &str) -> bool {
-    toc.lines().any(|l| {
-        l.split(" SCHEMA - ")
-            .nth(1)
-            .and_then(|rest| rest.split_whitespace().next())
-            == Some("public")
-    })
-}
-
-pub(crate) struct PreparedArchive {
-    path: PathBuf,
-    _tmp: Option<tempfile::TempDir>,
-    toc: String,
-}
-
-impl PreparedArchive {
-    pub(crate) fn path(&self) -> &Path {
-        &self.path
-    }
-    pub(crate) fn toc(&self) -> &str {
-        &self.toc
-    }
-}
-
-pub(crate) fn prepare_archive(
-    format: PostgresDumpFormat,
-    restore_file: &Path,
-    pg_restore: &Path,
-    logger: &JobLogger,
-) -> Result<PreparedArchive> {
-    let sniffed = crate::domain::postgres::connection::sniff_format(restore_file)?;
-    if sniffed != format {
-        logger.log("warn", format!("Declared format {:?} != sniffed {:?}; using sniffed", format, sniffed));
-    }
-    let format = sniffed;
-
-    let (path, tmp) = match format {
-        PostgresDumpFormat::Fc => (restore_file.to_path_buf(), None),
-        PostgresDumpFormat::Fd => {
-            let tar_gz = std::fs::File::open(restore_file)?;
-            let dec = flate2::read::GzDecoder::new(tar_gz);
-            let mut archive = tar::Archive::new(dec);
-            let tmp_dir = tempfile::TempDir::new()?;
-            archive.unpack(tmp_dir.path())?;
-
-            let dump_dir = if tmp_dir.path().join("toc.dat").exists() {
-                tmp_dir.path().to_path_buf()
-            } else {
-                std::fs::read_dir(tmp_dir.path())?
-                    .filter_map(|e| e.ok())
-                    .find(|entry| entry.path().join("toc.dat").exists())
-                    .map(|e| e.path())
-                    .ok_or_else(|| anyhow::anyhow!("Invalid FD archive: toc.dat not found"))?
-            };
-            (dump_dir, Some(tmp_dir))
-        }
-    };
-
-    let toc_out = Command::new(pg_restore).arg("-l").arg(&path).output()?;
-    if !toc_out.status.success() {
-        let stderr = String::from_utf8_lossy(&toc_out.stderr).to_string();
-        logger.log("error", format!("pg_restore -l failed: {}", stderr));
-        anyhow::bail!("Archive validation failed (pg_restore -l): {}", stderr);
-    }
-    let toc = String::from_utf8_lossy(&toc_out.stdout).to_string();
-
-    Ok(PreparedArchive { path, _tmp: tmp, toc })
-}
-
-pub(crate) fn run_pg_restore(
-    mut cmd: Command,
-    logger: &JobLogger,
-    cfg: &DatabaseConfig,
-) -> Result<()> {
-    let start = Instant::now();
-    let output = cmd.output();
-    let duration_ms = start.elapsed().as_millis() as f64;
-
-    match output {
-        Ok(o) => {
-            let stderr = String::from_utf8_lossy(&o.stderr).to_string();
-            let stdout = String::from_utf8_lossy(&o.stdout).to_string();
-            let combined = format!("{}{}", stdout, stderr);
-            let exit_code = o.status.code().unwrap_or(-1);
-            let payload = if combined.is_empty() { None } else { Some(combined) };
-
-            if o.status.success() {
-                logger.log_command("pg_restore", payload, Some(0), Some(duration_ms));
-                logger.log("info", format!("Restore completed successfully for {}", cfg.name));
-                Ok(())
-            } else {
-                logger.log_command("pg_restore", payload, Some(exit_code), Some(duration_ms));
-                logger.log("error", format!("Restore failed with status {:?} for {}", o.status, cfg.name));
-                anyhow::bail!("Postgres restore failed for {}", cfg.name);
-            }
-        }
-        Err(e) => {
-            logger.log_command("pg_restore", Some(e.to_string()), Some(-1), Some(duration_ms));
-            logger.log("error", format!("Error executing pg_restore for {}: {:?}", cfg.name, e));
-            Err(e.into())
-        }
-    }
-}
 
 pub async fn run(
     cfg: DatabaseConfig,
